payment method directive
