+++ src/js/controllers/buy-bitcoin/payment-methods.controller.js
@@ -14,7 +14,6 @@ angular
 
     // Functions
     vm.addPaymentMethod = addPaymentMethod;
-    vm.getIconPathFromName = getIconPathFromName;
     
     
     // Variables
@@ -25,17 +24,6 @@ angular
 
     function addPaymentMethod() {
       $state.go('tabs.buybitcoin-add-card-form');
-    }
-
-    function getIconPathFromName(name) {
-      switch(name.toUpperCase()) {
-        case "VISA": 
-          return "img/buy-bitcoin/icon-visa.svg"
-        case "MASTERCARD": 
-          return "img/buy-bitcoin/icon-mastercard.svg"
-        default:
-          return "img/buy-bitcoin/icon-generic-card.svg";
-      }
     }
 
     function _initVariables() {
