+++ www/views/includes/payment-method-item.html
@@ -1,0 +1,11 @@
+<div class="item item-icon-left item-big-icon-left">
+  <i class="icon big-icon-svg">
+    <img src="{{vm.getIconPathFromName(pm.brand)}}" />
+  </i>
+  <div class="apl-card__title">
+      {{pm.brand}} ••• {{pm.lastDigits}} 
+  </div>
+  <div class="apl-card__subtitle">
+    {{pm.expiryMonth}}/{{pm.expiryYear}}
+  </div>
+</div>
